Do some restructuring work
