+++ src/napari_imagej/_module_utils.py
@@ -10,8 +10,10 @@ from napari.layers import Layer
 from napari.utils._magicgui import find_viewer_ancestor
 from scyjava import JavaIterable, JavaMap, JavaSet, Priority, jstacktrace
 
-from napari_imagej._ptypes import TypeMappings, TypePlaceholders, _supported_styles
 from napari_imagej.setup_imagej import ij, jc, log_debug
+from napari_imagej.types.mappings import TypeMappings
+from napari_imagej.types.placeholders import TypePlaceholders
+from napari_imagej.types.styles import widget_for_item_and_type
 
 
 @lru_cache(maxsize=None)
@@ -610,29 +612,6 @@ def _add_param_metadata(metadata: dict, key: str, value: Any) -> None:
         pass
 
 
-def _widget_for_item_and_type(
-    item: "jc.ModuleItem",
-    type_hint: Union[type, str],
-) -> Optional[str]:
-    """
-    Convenience function for interacting with _supported_styles
-    :param item: The ModuleItem with a style
-    :param type_hint: The PYTHON type for the parameter
-    :return: The best widget type, if it is known
-    """
-    if type_hint == "napari.layers.Image" and item.isInput() and item.isOutput():
-        return "napari_imagej._helper_widgets.MutableOutputWidget"
-
-    style: str = item.getWidgetStyle()
-    if style not in _supported_styles:
-        return None
-    style_options = _supported_styles[style]
-    for k, v in style_options.items():
-        if issubclass(type_hint, k):
-            return v
-    return None
-
-
 def _add_scijava_metadata(
     unresolved_inputs: List["jc.ModuleItem"],
     type_hints: Dict[str, Union[str, type]],
@@ -654,7 +633,7 @@ def _add_scijava_metadata(
         if choices is not None and len(choices) > 0:
             _add_param_metadata(param_map, "choices", choices)
         # Convert supported SciJava styles to widget types.
-        widget_type = _widget_for_item_and_type(input, type_hints[input.getName()])
+        widget_type = widget_for_item_and_type(input, type_hints[input.getName()])
         if widget_type is not None:
             _add_param_metadata(param_map, "widget_type", widget_type)
