+++ tests/widgets/test_widget.py
@@ -1,106 +1,25 @@
-import pytest
 from napari import Viewer
 from qtpy.QtCore import Qt
-from qtpy.QtWidgets import QHBoxLayout, QLabel, QLineEdit, QVBoxLayout, QWidget
+from qtpy.QtWidgets import QApplication, QVBoxLayout
 
-from napari_imagej._flow_layout import FlowLayout
-from napari_imagej._helper_widgets import SearchEventWrapper
-from napari_imagej.setup_imagej import JavaClasses
 from napari_imagej.widgets.menu import NapariImageJMenu
-from napari_imagej.widgets.napari_imagej import (
-    FocusWidget,
-    NapariImageJ,
-    ResultTreeItem,
-    SearchbarWidget,
-    SearcherTreeItem,
-    SearchTree,
-)
-
-
-class JavaClassesTest(JavaClasses):
-    @JavaClasses.blocking_import
-    def ClassSearchResult(self):
-        return "org.scijava.search.classes.ClassSearchResult"
-
-
-jc = JavaClassesTest()
-
-
-def test_widget_layout(imagej_widget: NapariImageJ):
-    """Ensures a vertical widget layout."""
-    assert isinstance(imagej_widget.layout(), QVBoxLayout)
+from napari_imagej.widgets.napari_imagej import FocusWidget, NapariImageJ
+from napari_imagej.widgets.results import ResultsTree
+from napari_imagej.widgets.searchbar import ImageJSearchbar
+from tests.widgets.widget_utils import _populate_tree
 
 
 def test_widget_subwidget_layout(imagej_widget: NapariImageJ):
     """Tests the number and expected order of imagej_widget children"""
     subwidgets = imagej_widget.children()
     assert len(subwidgets) == 5
+    assert isinstance(imagej_widget.layout(), QVBoxLayout)
     assert isinstance(subwidgets[0], QVBoxLayout)
+
     assert isinstance(subwidgets[1], NapariImageJMenu)
-    assert isinstance(subwidgets[2], SearchbarWidget)
-    assert isinstance(subwidgets[3], SearchTree)
+    assert isinstance(subwidgets[2], ImageJSearchbar)
+    assert isinstance(subwidgets[3], ResultsTree)
     assert isinstance(subwidgets[4], FocusWidget)
-
-
-def test_searchbar_widget_layout(imagej_widget: NapariImageJ):
-    """Tests the number and expected order of search widget children"""
-    searchbar: SearchbarWidget = imagej_widget.findChild(SearchbarWidget)
-    subwidgets = searchbar.children()
-    assert len(subwidgets) == 2
-    assert isinstance(subwidgets[0], QHBoxLayout)
-    assert isinstance(subwidgets[1], QLineEdit)
-
-
-def test_results_widget_layout(imagej_widget: NapariImageJ):
-    """Tests the number and expected order of results widget children"""
-    results: SearchTree = imagej_widget.findChild(SearchTree)
-    assert results.columnCount() == 1
-    assert results.headerItem().text(0) == "Search"
-
-
-def test_focus_widget_layout(imagej_widget: NapariImageJ):
-    """Tests the number and expected order of focus widget children"""
-    focuser: FocusWidget = imagej_widget.findChild(FocusWidget)
-    subwidgets = focuser.children()
-    # Note: This is BEFORE any module is focused.
-    assert len(subwidgets) == 3
-    # The layout
-    assert isinstance(subwidgets[0], QVBoxLayout)
-    # The label describing the focused module
-    assert isinstance(subwidgets[1], QLabel)
-    # The button Container
-    assert isinstance(subwidgets[2], QWidget)
-    assert isinstance(subwidgets[2].layout(), FlowLayout)
-
-
-@pytest.fixture
-def example_info(ij):
-    return ij.module().getModuleById(
-        "command:net.imagej.ops.commands.filter.FrangiVesselness"
-    )
-
-
-def test_button_param_regression(ij, imagej_widget: NapariImageJ):
-    plugins = ij.get("org.scijava.plugin.PluginService")
-    searcher = plugins.getPlugin(jc.ModuleSearcher, jc.Searcher).createInstance()
-    ij.context().inject(searcher)
-    results = searcher.search("frangi", False)
-    assert len(results) == 1
-    searchService = ij.get("org.scijava.search.SearchService")
-    imagej_widget.focuser.focused_actions = searchService.actions(results[0])
-    py_actions = imagej_widget.focuser._actions_from_result(results[0])
-    assert py_actions[0].name == "Run"
-    assert (
-        imagej_widget.focuser.tooltips[py_actions[0][0]]
-        == "Runs functionality from a modal widget. Best for single executions"
-    )
-    assert py_actions[1].name == "Widget"
-    assert (
-        imagej_widget.focuser.tooltips[py_actions[1][0]]
-        == "Runs functionality from a napari widget. Useful for parameter sweeping"
-    )
-    assert py_actions[2].name == "Batch"
-    assert py_actions[2].name not in imagej_widget.focuser.tooltips
 
 
 def test_keymaps(make_napari_viewer, qtbot):
@@ -114,10 +33,8 @@ def test_keymaps(make_napari_viewer, qtbot):
     # Typing viewer.keymap['L'](viewer) does nothing. :(
 
 
-def test_result_single_click(make_napari_viewer, qtbot, asserter):
+def test_result_single_click(imagej_widget: NapariImageJ, qtbot, asserter):
     # Assert that there are initially no buttons
-    viewer: Viewer = make_napari_viewer()
-    imagej_widget: NapariImageJ = NapariImageJ(viewer)
     imagej_widget.results.wait_for_setup()
     assert len(imagej_widget.focuser.focused_action_buttons) == 0
     # Search something, then wait for the results to populate
@@ -144,109 +61,26 @@ def test_result_single_click(make_napari_viewer, qtbot, asserter):
     asserter(lambda: len(buttons) == 0 or all(button.isHidden() for button in buttons))
 
 
-class DummySearcher:
-    def __init__(self, title: str):
-        self._title = title
-
-    def search(self, text: str, fuzzy: bool):
-        pass
-
-    def title(self):
-        return self._title
-
-
-def test_searchers_disappear(imagej_widget: NapariImageJ, asserter):
-    # Wait for the searchers to be ready
-    tree = imagej_widget.results
-    tree.wait_for_setup()
-    # Update the Tree with some search results
-    searcher = DummySearcher("foo")
-    results = [jc.ClassSearchResult(c, "") for c in (jc.Float, jc.Double)]
-    tree.process.emit(SearchEventWrapper(searcher, results))
-    asserter(lambda: tree.topLevelItemCount() == 1)
-    asserter(lambda: tree.topLevelItem(0).childCount() == 2)
-
-    # Update the tree with no search results, ensure that the searcher disappears
-    tree.process.emit(SearchEventWrapper(searcher, []))
-    asserter(lambda: tree.topLevelItemCount() == 0)
-
-
-def _populate_tree(tree: SearchTree, asserter):
-    tree.wait_for_setup()
-    assert tree.topLevelItemCount() == 0
-    searcher1 = SearcherTreeItem(DummySearcher("Commands"))
-    searcher1.update(
-        [
-            ResultTreeItem(jc.ClassSearchResult(c, ""))
-            for c in (jc.Short, jc.Integer, jc.Long)
-        ]
-    )
-    tree.addTopLevelItem(searcher1)
-    searcher2 = SearcherTreeItem(DummySearcher("Ops"))
-    searcher2.update(
-        [ResultTreeItem(jc.ClassSearchResult(c, "")) for c in (jc.Float, jc.Double)]
-    )
-    tree.addTopLevelItem(searcher2)
-
-    # Wait for the tree to populate
-    asserter(lambda: tree.topLevelItemCount() == 2)
-    asserter(lambda: tree.topLevelItem(0).childCount() == 3)
-    asserter(lambda: tree.topLevelItem(1).childCount() == 2)
-
-
-def test_arrow_key_expansion(imagej_widget: NapariImageJ, qtbot, asserter):
-    # Wait for the searchers to be ready
-    imagej_widget.results.wait_for_setup()
-    # Search something
-    imagej_widget.results.search("Frangi")
-    tree = imagej_widget.results
-    asserter(lambda: tree.topLevelItemCount() > 0)
-    asserter(lambda: tree.topLevelItem(0).childCount() > 0)
-    tree.setCurrentItem(tree.topLevelItem(0))
-    expanded = tree.currentItem().isExpanded()
-    # Part 1: toggle with Enter
-    qtbot.keyPress(tree, Qt.Key_Return)
-    assert tree.currentItem().isExpanded() is not expanded
-    qtbot.keyPress(tree, Qt.Key_Return)
-    assert tree.currentItem().isExpanded() is expanded
-    # Part 2: test arrow keys
-    tree.currentItem().setExpanded(True)
-    # Part 2.1: Expanded + Left hides children
-    qtbot.keyPress(tree, Qt.Key_Left)
-    assert tree.currentItem().isExpanded() is False
-    # Part 2.2: Hidden + Right shows children
-    qtbot.keyPress(tree, Qt.Key_Right)
-    assert tree.currentItem().isExpanded() is True
-    # Part 2.3: Expanded + Right selects first child
-    parent = tree.currentItem()
-    qtbot.keyPress(tree, Qt.Key_Right)
-    asserter(lambda: tree.currentItem() is parent.child(0))
-    # Part 2.4: Child + Left returns to parent
-    qtbot.keyPress(tree, Qt.Key_Left)
-    asserter(lambda: tree.currentItem() is parent)
-
-
-def test_arrow_key_selection(imagej_widget: NapariImageJ, qtbot, asserter):
-    # Set up the tree
-    tree = imagej_widget.results
+def test_searchbar_results_transitions(imagej_widget: NapariImageJ, asserter, qtbot):
+    """
+    Ensures that the arrow keys can be used to transfer focus between
+    the searchbar and the results table
+    """
+    tree: ResultsTree = imagej_widget.results
     _populate_tree(tree, asserter)
 
-    # Ensure no row is selected
-    assert tree.currentItem() is None
-    # Go down, ensure that the first row gets highlighted
+    # Ensure that no element is highlighted to start out
+    asserter(lambda: tree.currentItem() is None)
+
+    # Then, press down from the search bar
+    # and ensure that the first item is highlighted
     qtbot.keyPress(imagej_widget.search.bar, Qt.Key_Down)
     asserter(lambda: tree.currentItem() is tree.topLevelItem(0))
-    # HACK: For some reason, they are not expanded in the tests!
-    tree.topLevelItem(0).setExpanded(True)
-    for i in range(3):
-        qtbot.keyPress(tree, Qt.Key_Down)
-        asserter(lambda: tree.currentItem() is tree.topLevelItem(0).child(i))
-    qtbot.keyPress(tree, Qt.Key_Down)
-    asserter(lambda: tree.currentItem() is tree.topLevelItem(1))
-    # HACK: For some reason, they are not expanded in the tests!
-    tree.topLevelItem(1).setExpanded(True)
-    for i in range(2):
-        qtbot.keyPress(tree, Qt.Key_Down)
-        asserter(lambda: tree.currentItem() is tree.topLevelItem(1).child(i))
-    qtbot.keyPress(tree, Qt.Key_Down)
-    asserter(lambda: tree.currentItem() is tree.topLevelItem(1).child(1))
+
+    # Then, press up and ensure that the search bar is highlighted
+    qtbot.keyPress(tree, Qt.Key_Up)
+    # HACK: In practice, we should expect focus on imagej_widget.search.bar.
+    # But because it isn't visible, QApplication.focusWidget() will become None.
+    # See https://doc.qt.io/qt-5/qwidget.html#setFocus
+    asserter(lambda: not imagej_widget.search.bar.isVisible())
+    asserter(lambda: QApplication.focusWidget() is None)
